include/avisynth_c.h: "1 << 31" raises a warning on C99/C11/C17
Avoid overflow in C99/C11/C17.

+++ include/avisynth_c.h
@@ -120,7 +120,7 @@ enum {
     AVS_CS_BGR = 1 << 28,
     AVS_CS_YUV = 1 << 29,
     AVS_CS_INTERLEAVED = 1 << 30,
-    AVS_CS_PLANAR = 1 << 31,
+    AVS_CS_PLANAR = (int)(1u << 31),
 
     AVS_CS_SHIFT_SUB_WIDTH = 0,
     AVS_CS_SHIFT_SUB_HEIGHT = 8,
